HQ-326 Moodle + Jenkins init script
This scripts perform all the actions needed to get one
default jenkins server ready to work with moodle ci checks
creating all the directories, databases, database users and
the moodle ci site instance.

It's safe to execute it multiple times as far as every action
is verified before execution.

Take a look to the comments in the script to know the requirements
and the environmental variables needed.

It can be executed manually or from within Jenkins. The later, the
recommended. Take a look to the "Initial Jenkins Moodle Setup" job
in Moodle integration ci servers.