MDLSITE-2005 Waiting for Inte => Reopened ignored
Take out the Waiting for Integration => Reopened transition
because it was leading to a lot of false positives being
counted. It's really a pitty not being able to specify, in
historic searches, any custom field, like "Currently in integration".

So, once applied, this job will look, exclusively for these
"transitions" since last integration cycle was closed:
  - Integration in progress => Reopened
  - Waiting for testing => Reopened
  - Passed => Reopened
  - Failed => Reopened
  - Tested => Reopened