MDLSITE-2100 ci: never reset to origin, can contain newer commits
All the jobs that are chained per branch cannot be reset to
their origin/xxxx counterpart because it's possible that the
git poll already had incorporated commits for the next run. So
the reset is performed to local branch and done.

Note that the git poll itself is configured to renew the local
branches once the next run starts (not before, poll knows that
it has to wait until the whole chain ends before renewing local
branches).