MDL-38904 Remove exclusion for bootstrap js
Bootstrap uses a yui gallery module with a single whitespace error.
The module is no longer maintained so is not likely to be updated -
since MDL-38904 moved the js and I would have had to update CI anyway,
I just fixed the whitespace error and am removing the exclusion from ci.