NOBUG: ignores now used both by checker and setter
The list of version.php files to ignore was harcoded
both in the checker and the setter and, when executing
the setter @ MDL-39434 it was detected that the list
had become out of sync.

This change just unifies the list of ignores (regular
expression under $gitdir) and applies it both to the
checker and the setter.