NOBUG: Run the script against installed site.
Before this change, the script was being executed
from the moodle_ci_site by hacking CFG->dirroot and
reseting caches, so plugin information was gathered
from an uninstalled site.

After this change, the shell script creates a real
site and the php script is executed from that just
created site. So the list of plugins is 100% real.

Is slower, but the only way to get it 100% accurate.