NOBUG: Once the test has run, delete local/ci
Since 2.6 version.php files are expected to be present
for all the plugins installed. This test copies some scripts
to gitdir/local/ci in order to execute stuff with the site
installed using that gitdir.

Once the test ends, we need to delete local/ci or phpunit
tests will detect it as a "wrong" (missing version.php) plugin.