NOBUG: Split processing server url and publish one.
It has been detected that pointing to the public URLs of the
jenkins servers leads to stream problems when the jenkins-cli
is controlling the execution of jobs. Instead it must use
the "direct" url so communications aren't broken. In the other side,
the publish URL is needed to spread results out there, allowing any
developer to connect and watch results.

Also integration_held issues have been excluded from both the
peer-review and awaiting-integration lists.