MDLSITE-2714 All criteria to clean the cime label
Before this, it was possible for the chain of precheckers
to pick the same issue more than once, if the "cime" label
was added to an issue sent to peer review or integration.

With this change, he "cime" label is always cleaned by all
criteria so no dupe executions are possible anymore.