MDLSITE-2882 look for exceptional stuff in output
Apart from looking for any exception, debugging, stack trace
along the execution of the script, we are also looking for
any information in output not being part of the standard
progress report (lines of dots with some numbers and % at the end).

This should allow us to detect anything incorrect, sent
with error_log() and friends (both to standard and error output).