NOBUG: Handle moodle.git clones internally.
With this change the job does not need the
presence of the git plugin defined anymore.

Right now it was more an obstacle than a help
because it had to be configured strangely (pointing
to a old branch/commit), to avoid it leading to
the job own operations in the git repo.

So, with this change the job itself manages both the
initial clone and the cleaning before proceeding with the checks.