MDLSITE-1990 ci: enforce checkstyle generation.
Before this, both the savepoints and the commit checkers were
generating and empty .xml checkstyle file. In the other side,
both the code checker and the phpdocs checker were always
generating their .xml files.

That caused the former ones to not be present in the final smurf
files always, making it hard to use a constant checksum to look
for perfect issues (no warnings and no errors).

So, now... those .xml files will be always created, no matter they
don't contain any problem.