MDLSITE-2712 tidying, commenting and reorganizing a bit
This change contains basically cosmetic changes only to
comments, initial creation of integration.git remote and
tidying exit-on-error.

The only remarkable change is that, now, the ancestor to be
used is calculated (using the same way it was previously) and
only then all the checks begin (including the branch aging ones
that previously were executed always for the moodle.git ancestor).