MDLSITE-2012 prechecker: Detect plugin builds and adapt checks
Agreed with David, all his builds from the Plugins database will
be passing PLUGIN-xxxxx (xxxx being the plugin version id) to the
prechecker.

That way the prechecker will know the execution is against a plugin
and will skip the commits checker (for now) and will be able to
introduce other customizations along the time (keeping the prechecker
working for everybody).