CONTRIB-4696 Apply distributed thirdpartylibs.xml to CI jobs
Originally we introduced the ability to automatically exclude
all the thirdpartylibs.xml locations to the local_codechecker
plugin (phpcs invoked from within Moodle).

CI jobs do not use that phpcs but the stock one that is not
aware of our thirdpartylibs.xml at all. Instead, it relies in
a fixed list of exclusions, that has worked nicely till now
for core.

Recently we have started to use the CI jobs to test plugins
that can come with their own thirdpartylibs.xml, and those
exclusions are not being applied.

This patch just looks for all the available thirdpartylibs.xml
files and add as many <location> exclusions are defined there.