MDLSITE-3769 local_ci: Automated the count of delayed issues.
As far as it's not possible to look for comment date (grrr), we
look for all issues that have been waiting for integration or
integration in progress and did not transition beyond that and
have the delayed comment on them.

Not sure if this is going to work or we'll need to fetch comments
and filter by date manually... but worth trying as far as doing
the later can be really slow.