MDLSITE-4023 bulk prechecker: Look for controlled errors.
We output to the errors.txt file information with different
levels (Info, Warn, Error). Before this patch, any level was
causing the errors.txt to be printed, leading to confusion.

Now we check if there is any Error level message to print that
file. If there isn't any error, then an uncontrolled error has
happened and we should be warned in order to fix it.