MDLSITE-3874 ci: make source per branch to use node_modules subdir
For some reason, soft linked installation of node stuff does not
work well unless the source directory is called node_modules. Sure
there are some realpath bit somewhere (grunt?) depending of that.

Ultimately it was causing grunt to fail, not being able to execute
the uglify-js task. With this trick, everything seems to work ok.