NOBUG: Avoid weird dereferenced links std behavior.
By default, ln itself follows links so, at the end
it's highly possible to get them recursively created
no matter they already exists. To solve this the -n
flag can be used, causing ln not to follow links.

Or, something like that, really I've not thought too
much about it. I just tried the flag and verified it
solved the problem, so applying it everywhere.

Ref: http://www.gossamer-threads.com/lists/linux/kernel/1966