NOBUG: Allow the maxcommits to be configurable
While we recently raised maxcommits to 100 (harcoded in the script),
there are situations where we want to raise the limits more,
specially when we are importing a branch containing years-of-dev
about a new plugin.

In those situations it's interesting to launch the bulk-checker manually
by incrementing the number of allowed commits.

A real example: https://tracker.moodle.org/browse/MDL-47494
(there you can see all the "failed" CiBoT attempts and the interest
into being able to, exceptionally, run the checker with more commits
allowed).