MDLSITE-4198 ci: split maxcommits into warning(10) and error(100)
To better support branches with a number of commits, the old, unique,
maxcommits setting is changed plugin-wide to use a couple of settings
instead.

Also, we are moving from an eroor&stop behavior (where the rest of the
checks were not applied) to a error/warn and continue behavior, by
showing the maxcommits warnings and errors within the commits checker.

This may lead to problems is somebody does submit a huge/wrong branch,
but that also was happening right now irrespectively of the # of
commits. If we need countermeasures for that situation, time will tell.