MDLSITE-3818 cleanup: kill BUILD_ID and delete issues details
This commit cleans all remaining jobs using BUILD_ID and, instead
do use a calculated (now), BUILD_TIMESTAMP, that is what BUILD_ID
used to contain. Note this is basically a cosmetic change to make
scripts to show information as it was some time ago.

Also, in various tracker jobs where a number of issues (with full
details) are returned, to avoid any disclosure of (security...)
information, the file where details are stored is explicitly
deleted at the end of the process.