MDLSITE-4366 savepoints: discard unsafe strings to avoid problems
Now we pre-process the upgrade code, looking for all strings
within it. If the string is considered dangerous (contains
parenthesis and other chars, then it's replaced by safe placeholder
so next regexp operations won't fail).

Safe strings are kept in place in order to keep code transformations
being as small as possible, in case we want to add more checks in
the future (savepoint params checks...).