NOBUG: Centralize all the npm install stuff
Right now there are a bunch of jobs requiring the
installation of different nodejs stuff with npm.

A) Chain of jobs by branch.

- For branches < 29 (without package.json file):
  - The job 'shifter_walk' uses 'shifterbase' to maintain different
    shifter versions per branch. (current value ~/shifter_base)
  - The job 'less_checker' uses 'recessbase' to maintain different
    recess versions per branch. (current value ~/recess_base)
- For branches >= 29 (with package.json file):
  - The job 'grunt_process' uses 'npmbase' to maintain different stuff
    (grunt, shifter, less, jshint...) per branch. (current value
    ~npm_base).

B) Prechecker.

  - The job 'remote_branch_checker'' uses 'npmbase' to maintain different
    stuff (grunt, shifter, less, jshint...) per branch. (current value
    ~npm_base_prechecker). Run by 'grunt_process' above.

C) New uses could also require the installation of npm stuff. For sure
   in a different base from the already used above.

So, this new job is a "universal" nodejs stuff installer, behaving as
follows:

1) npm will be in charge of installing any package.
2) every use above (A, B, C) will use a different 'npmbase'.
3) within every 'npmbase' stuff will be installed per branch, aka:
  - A's own npmbase/branchname/node_modules
  - B's own npmbase/branchname/node_modules
  - ....
4) If package.json is available, both grunt-cli and package.json stuff
   will be installed. This is the way in 29 and up.
5) If package.json is not available, but shifterversion has been
   specified, shifter@version will be installed. This is for < 29.
5) If package.json is not available, but recessversion has been
   specified, recess@version will be installed. This is for < 29.
6) At the end of the process all the installed binaries @
   node_modules/bin will be listed.
7) Link the correct node_modules directory in the gitdir being used. The
   job using the installer will be in charge, if needed, of deleting the
   link once the job is done.

Important note: This job does not perform any git operation (fetch,
checkout...). Just requires the gitdir and gitbranch params in order to
decide where packages will be installed and linked to.

Once this job is available, the idea is to use it for new uses (C above)
and also, progresively, replace all the current A & B uses above.

That's it. Next step, replace, in the A, B jobs above, current partial
installers of nodejs stuff by this universal alternative.