MDLSITE-4702: Without moodle default chdir() we need to use absolute paths
With PR #109 we deleted the job dependencies on Moodle. But it seems
we forgot to notice (in fact tests were passing) the automatic chdir()
call present in setup.php. Without it, no xlst sheets were found and
the job ultimatelly was failing. Simple fix does switch to full paths
so xslt sheets will be found without problem.