MDLSITE-4702: Add local helpers for doing running of scripts
These helpers change the run behaviour so we no longer run the
scripts in the same directory as the script as previously. The previous
behaviour doesn't match the way which scripts are run with in the jenkins
environement and can lead to errors with paths not included safely.

This should hopefully prevent a repeat of db90151cae6a28