MDLSITE-4767: Fix regexp variant of excluded to use proper regexp
Apparently it was using some sort of * shell expansion, hence not
working always. Now * is translated to proper .* regexp.

As far as now *.csv are being excluded, we don't need individual
lib/tests/fixtures/tabfile.csv anymore.

Also get rid of mega-old "search" exclude that was causing lots of
new search stuff not to be checked.

And, finally, also "node_modules" that normally CIs don't face, but
can be a source of problems/confusion when working locally.