MDLSITE-4790 mustache_lint: make eslint less noisy
Based on some internal discussions I convinced myself:

1) Not to output parse errors when there is missing example context,
because this is often caused by the missing example context, its just
adding noise which is covered by the previous warning.

2) We should make all problems warnings because again they might be
caused by missing example context. This is the same way we treat html
validation

If we had a codebase without these missing example context problems this
wouldn't be an issue.