MDLSITE-4806 remove the "ci" label when issues arrive to integration
When an issue, having the ci label arrives to "waiting for integration"
it's interesting to get the label cleaned, so the prechecker can verify
that everything is correct against current upstream.

This is specially noticeable in continuous-integration periods where
the issues are never moved to "current integration" (monday automated
job) so last prechecker results are really old (happened normally before
peer-review).

This job picks those issues and cleans the label, forcing an extra
prechecker to happen. Has built in machinery to prevent the same issue
to be cleaned too many times (in 1 hour) and other details. Should be
safe to be executed at any time interval.