NOBUG: Always delete any _precheck branch from previous executions
We cannot delete the _precheck branches while they run because that
will cause all the /work to be lost before being saved as artifacts.

So, on every run, before performing any other git operation, we check
if we are in a _precheck branch and, if so, proceed to checkout its
base counterpart, checkout it and delete the _precheck branch and
any old stuff it may have.

Note this was introduced because we found some plugins having own
.gitattributes files, playing with auto-crlf settings and causing
next executions of the checker to fail forever. Now, we ensure
nothing is left in the prechecker repo from previous executions.