MDLSITE-4892 deal with spaces in passed varibles
Because this comes from tracker fields, we can trim the spaces. Probably
would be more appropiate in the remote branch checker, but its better
covered with tests here and doesn't do any harm.

I decided not to add a testcase for the 'ok with space' case to avoid
wasting the time of a full precheck run.