MDLSITE-5184 mustache_lint: Ensure eslint uses correct .eslintrc
I was getting a consistent (testing) failure with error:

Failed to load plugin promise: Cannot find module 'eslint-plugin-promise'

And it was incredible because the test is using MOODLE_31_STABLE branch
and the promise plugin was not used by that branch. At the end, it seems
that eslint somehow merges the configs passed in the call and the
configs present in the path (down-top). As far as I was running local_ci
within a master moodle site, the eslint "promise" plugin was added to
the execution and failed because the 31_STABLE node_modules does not
have it.

By adding --no-eslintrc we prevent eslint to search in own path, so only
the --config alternative is considered. That's what we want. Exactly.