NOBUG: Parametrise jenkins cli connectiviy string vs prev. harcoded one.
New versions of Jenkins CLI ask you to disable old, insecure CLI
protocols in the Jenkins server. Because of that some "magic"
authentication that was happening (jenkins user) does not work
anymore. So we need to be able to pass the required connectivity
setting to newer jennkins cli, more exactly:

-s http://jenkis_public_url:8080 -ssh -user clibot

to specify that we want to use localhost, via jenkins-own-ssh
and with the keys of the jenkins user "clibot" that are the keys of
os user "jenkins" running the jira client in shell. So they work.

Note that the Jenkins own ssh port must be accessible by the caller
so using -Dorg.jenkinsci.main.modules.sshd.SSHD.hostName can be handy
if the public Jenkins URL is behind proxy or whatever. In general all
the local_ci uses are local, so it's ok to have:

-Dorg.jenkinsci.main.modules.sshd.SSHD.hostName=localhost

and ignore any other complication or need to make jenkis_public_url
available for ssh purposes.