NOBUG: Grunt to ignore changes caused by 3rd part libs
Some plugins can come with their own 3rd part libs (thirdpartylibs.xml).
That leads to grunt feeding .eslintignore and .stylelintignore to
make js & css linters to ignore those 3rd part libe.

But those changes, make the grunt check to fail as far as they lead
to change in core detected when the check is being performed for a
non-core plugin. So we just proceed to ignore changes in those files
when it's the case.