MDL-63346 npm: Moving away from git pathspecs to simpler grep excludes
For some reason the git pathspec excludes aren't working as expected
when run in subshell (maybe it's some weird escaping, expansion...) and
haven't been able to keep it working for multiple.

So, switching to old known filter out of results using grep -v
and done. It supports multiples without problm and just works.