MDLSITE-5313 make jenkins-cli stop consuming stdin
In the context of a "while read..." loop, it was consuming
all the lines silently, causing the loop to aparent that
only the fist line was processed.

Note that, for regular ssh, there is a (-n) switch that produces
exactly the same effect. But here, it's used internally by jenkins-cli
and there isn't any way to passthrough those switches. Or tomorrow we
could be changing to any other connection protocol offered by Jenkins.

So < /dev/null seems safe (there isn't anything more to read from stdin)
and will continue working for any underlying stuff used within
jenkins-cli.