Launch predefined jobs against issues returned by a query
This job just allows you to perform JQL queries against
a Jira server, launching a number of defined jobs for each
issue found and performins some post-actions for each issue.

We are going to use it with a criteria named "awaiting_integration"
so, every user that lands to integration gets some CI jobs executed
for every branch in the issue.

At Jenkins level we'll be throttling those (DEV) jobs to prevent
this script to end using all the available workers.

It's important to make the schedulemins parameter to match the
intervals (in minutes) in the job schedule plan (cron) or some
issues could be processed twice or skipped.

Finally note that the current criteria includes a extra condition:

AND participants not in (tobic)

So we'll be processing an issue only once. This can be tidied/chaged
later, but for now it's a basic protection trying to avoid an issue
going back and forth to consume too much infrastructure resources.

Also note that, for now, we have disabled the Oracle PHPUnit job, just
to sabe some workers until we have the job controlled and decide that
we can add more jobs to it.

This job will be launched by the "TR - Bulk pre-launch DEV jobs
(awaiting_integration)" job that has been already created and is
waiting for this to land to perform the final configuration and
schedule it.