Tidy the search to verify the status ON old date, not BEFORE
Previously it was only returning issues that never had been
in 'Waiting for integration review' status, keeping out
issues thay may have been already under integration.

Now we just look for issues that changed from anything in
the past date to 'Waiting for integration review'.

Note we cannot use the ideal CHANGED TO ... because there
are a bunch of transitions that change from waiting to waiting
and they are returned. We don't want them.

For normal executions at fixed intervals (2h right now) it won't
make much difference in practice, so all right