Automatically follow .nvmrc when available
This enables all jobs using the script (security rebase, prechecker,
grunt, ... to, automatically use the current node version speficied
in the .nvmrc file.

This includes:
- installing nvm if not available ($HOME/.nvm)
- update nvm to the latest release
- source nvm
- install and use the .nvmrc version

Note that, once working, this renders the use of the nvm-wrapper
jenkins extension futile, as far as it does everything the extension
was doing previously. But better, because now we can control when
our moodle directoy is ready and with the correct .nvmrc file.

This will help to issues like MDL-66109, so whenever a node bump
happens, it's automatically detected and handled by our scripts
that will bump and use the new node version transparently.

If there is no .nvmrc file, nvm is not sourced, installed or used.
Assumed that somebody else has already prepared all the stuff.