Avoid sourced script to overwite caller $mydir variable
Important, this script is always run sourced from others, so it shouldn't define
any shell variable also set/used by caller scripts. A clear example is the $mydir
variable below, that was being set for caller script, leading to strange failures.

In this case it was easy to fix, because this script doesn't use it, so getting rid
of it was enough. But in general, avoid setting any widely used variable here.