Fix change of behavior in eslint v6.8 and up
Old eslint versions, when there weren't any JS file to inspect
just exited silently and without error (exit status = 0).

But since eslint v6.8.0, if there aren't JS files, it exits
with error (exit status = 1). A new option,
--no-error-on-unmatched-pattern, has been created to provide
the BC bahavior.

So we need to discern which eslint is being used. Initally I
did it looking for --version... but it's inexact.

So, surely, the safest way is to verify if the new option exists
or no (using --help). And decide based on that.

That's exactly what this patch does. I've also added a TODO
because one everybody is already using nodejs 14 we won't need
this hack anymore.

Also, added pretty much the same for stylelint, but using the
--allow-empty-input option.

Finally note that there is another eslint use (within the mustache
linter) that is not fixed here. It will come in a new PR because
the fix is different (eslint is executed from within PHP).