Implement the interval checker to ensure all versions are within limits
Tested against all branches, it seems that we have a couple of
violations in 35_STABLE and 38_STABLE. Unlikely to cause problems
but glad the tests detected them.

We are going to enable these checks for 39_STABLE and up, using
the following allowed version intervals:

3.5: 20180514-20180517 (grrr there is a 2018061801 version!)
3.7: 20190520
3.8: 20191118 (grrr there is a 2020031700 version!)
3.9: 20200615
3.10: 20200615-20201109
master: 20210525

Will be also modifying the release process to add a point about
keeping the intervals updated when releases happen.