Add possibility to exclude selected Mustache files from HTML validation
When developing for the Moodle Mobile App, the Mustache templates may
contain Ionic3 / Angular specific syntax that itself is not valid HTML.
The patch adds an option to skip the HTML validation part of the
Mustache linting.

To ignore HTML validation errors, create a file .mustachelintignore in
the directory with Mustache templates and define the file name patterns
for the files to be excluded from HTML validation. PHP function
fnmatch() is used to search for matches. Lines starting with the hash
sign are considered as comments.