Stop requiring Travis
Most people have moved away from Travis so we should stop warning if
it is not used.

THis change modifies the Travis integration tests to mark the test as
SKIPped if the integration is not setup. A warning is still used if
travis-ci.org is configured but not travis-ci.com.