Allow to decide the target for the ToBiC comments
Right now, all the comments in the criteria are
harcoded to Integrators.

This change allow you to pick any other role in the project
with blank meaning visible to all.

The jobs @ ci.moodle.org has been already modified to support
this new parameter that will be configurable when any of the
jobs there are executed manually. By default, they continue
restricting to Integrators, until we decide to open them.