Add some exceptions within tool/brickfield
All the cases are sort of test/fixture files and license. So maybe
it's not crazy to review them and, if some can be cleaned, proceed.

So https://tracker.moodle.org/browse/MDL-71532 has been created
about that. Once fixed, surely we'll be able to remove some
of the exceptions being added here.