Prevent scripts against some buggy JAVA en_AU locales
From CLDR29 to CLDR33, the month and day names for en_AU were broken, incorrectly
adding a dot (.) after the day and month names. And it was fixed in CLDR34 . So I imagine
that java versions >=9 (when they changed from own locales to CLDR) that are using those CLDR
versions… are doomed (that for sure includes Java10).

Alternatively, there is the following switch:

-Djava.locale.providers=COMPAT, CLDR

to make Java own locales to be used instead of CLDR ones.

But as far as en_AU is already fixed in actual versions, we are simply
removing those incorrect dots fom the date fetched with Jira CLI and
done.

Some links:

- https://bugs.openjdk.java.net/browse/JDK-8208487