Avoid the normal queues manager to play with blocks/is blocked by issues
The set_integration_priority_to_[zero|one] scripts are in charge of
controlling the integration priority of both blockers and blocked issues
(see the scripts for more details), so we don't want the normal queues
manager to perform any change of priority on them.

Note that this is the only other actually active script playing with
integration priority (there is also the old delay_awaiting_issues one
but we don't use it anymore because was replaced by the queue managers).