Verify that applied branches and fix-versions in the tracker match
Till now just searched by unowned commits, without ever matching
if the fix-versions effectively match the branches that have received
the fix.

Now, given the list of fix-versions for an issue, we calculate
the expected branches where it should be and verify if any of
the expectations is missing. Because, here and there, we forgot
to add some stable (normally) fix version and the job continued
passing. Now it will ask to fill the fix-versions properly