Always verify if the candidate issues are blocked by others
Previously it only was checked for "blocker" issues (one of
the conditions to consider the issue important). Now it's checked
as for all the conditions:
- mldqa
- must-fix
- security
- blocker

Old behaviour was causing some ping-pong with issues being raised
and immediately after that, being lowered. Now, they won't be
raised ever if it's blocked by any unresolved.