Pass the complete list of components to phpcs
Normally, phpcs itself is able to get all the components
installed in a given dir root (it just loads them using
standard core_component class).

But, for performance reasons, with CiBoT, we delete all
files from dir root but those being part of the patchest
being analysed (and a few exceptions like version.php, grunt...).

That means that it's unable to load core_component or find all
the components available in a site.

But that's ok, because one of the first things that the remote
branch checker already does (before deleting files), is to, always,
calculate all the valid components and save them to the
/work/valid_components.txt file.

And then, codechecker can use the new config setting
"moodleComponentsListPath" to load exactly that file. See
https://github.com/moodlehq/moodle-local_codechecker/pull/160
for more information about that feature.

Covered with tests to verify that it's getting the components
information properly.