Avoid blocked issues to "consume" to be moved seats
Before the patch, every blocked issue was being accounted
as a move (because before we implemented the blocked detectors
all issues were, unconditionally, moved).

Now, we account to really moved issues in another ($moved)
variable and allow the process to continue until we have
really moved $movemax issues.

This way, blocked issues (or any other condition/check) that
we may want to introduce in the future won't affect the outcome
of moving up to $movemax issues.