Change the query used to count reopened issues
Previously it was looking for FROM...TO transitions but it's
possible that the transition happens from other statuses.

So, now, we count ANY issue that was under integration in process,
waiting for feedback or testing failed statuses and has been reopened.

Both conditions since the last time a roll happened, of course (when
the gates were open to continue integrating).