After release, first held issue, then keep the queue fed
We have just swapped the steps, because, after release (while on-sync)
we were first moving the issues to current and later looking for
new improvements and features to be held (and none was being found,
obviously because they had been moved).

So this change ensures that, after release (it was correct before
release). We first held improvements and new features (B1a) and
only then consider the remaining issues to keep the current
integration fed (B1b).