Create the phpunit.xml file not requiring database or real site
In order to create the phpunit.xml file, PHPUnit requires a
real database / site to be able to do its job. But here we are
@ CIs and don't want to spend resources and time doing the
whole phpunit boostrap (database connection, composer stuff...).

So this CLI just workarounds all those requirements and straight
calls to phpunit_util::build_config_file() that is able to
do all the work with just a few includes.

This uses the also, moodle-less load_core_component_from_moodle()
in order to get all the components loaded.